[all] rethink error handling with Context
This removes the custom MissingRequired errors.
