+++ cli/src/lib.rs
@@ -29,3 +29,5 @@ pub mod ops;
 pub mod config;
 pub mod errors;
 pub mod output;
+mod build_spec;
+mod config_env;
